backend: Restructure code by role
Backend responses that are specific to each role should be placed under their respective role, with common responses given their own package in common
