+++ backend/src/main/java/com/cs471/studentLoanSystem/common/login/response/RegistrarResponse.java
@@ -1,0 +1,3 @@
+package com.cs471.studentLoanSystem.common.login.response;
+
+public class RegistrarResponse extends LoginResponse {}
