--- backend/src/main/java/com/cs471/studentLoanSystem/login/response/RegistrarResponse.java
@@ -1,3 +1,0 @@
-package com.cs471.studentLoanSystem.login.response;
-
-public class RegistrarResponse extends LoginResponse {}
